requirements.txt
# Runtime + library (Milestone data-builder)
httpx>=0.27
playwright>=1.49.0
beautifulsoup4>=4.12
lxml>=5.0
pydantic>=2
pydantic-settings>=2
python-dotenv>=1.0
supabase>=2.10.0
anthropic>=0.40.0
googlemaps>=4.10.0
pandas>=2.2.0
openpyxl>=3.1.0
thefuzz>=0.22.0
geopandas>=0.14.0
shapely>=2.0.0
requests>=2.31.0
apscheduler>=3.10.0
pytz>=2024.1
# Editable local package (data_builder config + shared utilities)
-e .
